fix(sideNavBar): demo拆解与规范
