+++ src/packages/sidenavbar/doc.en-US.md
@@ -16,82 +16,7 @@ import { SideNavBar,SubSideNavBar,SideNavBarItem } from '@nutui/nutui-react';
 
 :::demo
 
-```tsx
-import React, { useState } from 'react'
-import { Cell, SideNavBar, SubSideNavBar, SideNavBarItem, Toast } from '@nutui/nutui-react'
-
-const App = () => {
-  type Position = 'left' | 'right'
-  type NavBarState = {
-    visible: boolean
-    position: Position
-  }
-  const [navBarState, setNavBarState] = useState<NavBarState>({
-    visible: false,
-    position: 'left',
-  })
-  const [showThird, setShowThird] = useState(false)
-  const changeNarBar = (visible: boolean, position: Position = navBarState.position) => {
-    setNavBarState({
-      visible,
-      position,
-    })
-    setShowThird(false)
-  }
-  const clickItem = (data: any) => {
-    const { title, value } = data
-    showThird && Toast.show(`title=${title},value=${value}`)
-  }
-  const clickTitle = (data: any) => {
-    const { title, value, isShow } = data
-    showThird && Toast.show(`title=${title},value=${value},isShow=${isShow}`)
-  }
-  return (
-    <>
-      <Cell
-        title='Left Pop Up'
-        onClick={() => {
-          changeNarBar(true, 'left')
-        }}
-      />
-      <Cell
-        title='Right Pop Up'
-        onClick={() => {
-          changeNarBar(true, 'right')
-        }}
-      />
-      <SideNavBar
-        title='Front Page'
-        visible={navBarState.visible}
-        position={navBarState.position}
-        onClose={() => {
-          changeNarBar(false)
-        }}>
-        <SubSideNavBar title='Level One Title' value='1-0' onClick={clickTitle}>
-          <SideNavBarItem title='Level One Content1' value='1-01' onClick={clickItem} />
-          <SideNavBarItem title='Level One Content2' value='1-02' />
-          <SubSideNavBar title='Level Two Title' value='2-0'>
-            <SideNavBarItem title='Level Two Content1' value='2-01' />
-            <SideNavBarItem title='Level Two Content2' value='2-02' />
-            {showThird ? (
-              <SubSideNavBar title='Level Three Title' value='3-0'>
-                <SideNavBarItem title='Level Three Content1' value='3-01' />
-                <SideNavBarItem title='Level Three Content2' value='3-02' />
-              </SubSideNavBar>
-            ) : null}
-          </SubSideNavBar>
-        </SubSideNavBar>
-        <SubSideNavBar open={false} title='Level One Title-2' value='1-1'>
-          <SideNavBarItem title='Level One Content2-1' value='1-11' />
-          <SideNavBarItem title='Level One Content2-2' value='1-12' />
-        </SubSideNavBar>
-      </SideNavBar>
-    </>
-  )
-}
-export default App
-
-```
+<CodeBlock src='h5/demo1.tsx'></CodeBlock>
 
 :::
 
@@ -99,76 +24,7 @@ export default App
 
 :::demo
 
-```tsx
-import React, { useState } from 'react'
-import { Cell, SideNavBar, SubSideNavBar, SideNavBarItem, Toast } from '@nutui/nutui-react'
-
-const App = () => {
-  type Position = 'left' | 'right'
-  type NavBarState = {
-    visible: boolean
-    position: Position
-  }
-  const [navBarState, setNavBarState] = useState<NavBarState>({
-    visible: false,
-    position: 'left',
-  })
-  const [showThird, setShowThird] = useState(false)
-  const changeNarBar = (visible: boolean, position: Position = navBarState.position) => {
-    setNavBarState({
-      visible,
-      position,
-    })
-    setShowThird(false)
-  }
-  const clickItem = (data: any) => {
-    const { title, value } = data
-    showThird && Toast.show(`title=${title},value=${value}`)
-  }
-  const clickTitle = (data: any) => {
-    const { title, value, isShow } = data
-    showThird && Toast.show(`title=${title},value=${value},isShow=${isShow}`)
-  }
-  return ( 
-    <>  
-      <Cell
-        title="Show"
-        onClick={() => {
-          changeNarBar(true)
-        }}
-      /> 
-      <SideNavBar
-        title='Front Page'
-        visible={navBarState.visible}
-        position={navBarState.position}
-        onClose={() => {
-          changeNarBar(false)
-        }}>
-        <SubSideNavBar title='Level One Title' value='1-0' onClick={clickTitle}>
-          <SideNavBarItem title='Level One Content1' value='1-01' onClick={clickItem} />
-          <SideNavBarItem title='Level One Content2' value='1-02' />
-          <SubSideNavBar title='Level Two Title' value='2-0'>
-            <SideNavBarItem title='Level Two Content1' value='2-01' />
-            <SideNavBarItem title='Level Two Content2' value='2-02' />
-            {showThird ? (
-              <SubSideNavBar title='Level Three Title' value='3-0'>
-                <SideNavBarItem title='Level Three Content1' value='3-01' />
-                <SideNavBarItem title='Level Three Content2' value='3-02' />
-              </SubSideNavBar>
-            ) : null}
-          </SubSideNavBar>
-        </SubSideNavBar>
-        <SubSideNavBar open={false} title='Level OneTitle-2' value='1-1'>
-          <SideNavBarItem title='Level One Content2-1' value='1-11' />
-          <SideNavBarItem title='Level One Content2-2' value='1-12' />
-        </SubSideNavBar>
-      </SideNavBar>
-    </>
-  );
-};  
-export default App;
-
-```
+<CodeBlock src='h5/demo2.tsx'></CodeBlock>
 
 :::
